Extract endpoint details store
